SecureAuthentication (2)/SecureAuthentication/requirements.txt
Flask==2.3.3
flask-cors
flask-sqlalchemy
opencv-python
deepface
numpy
pillow
